+++ hackru-frontend/public/index.css
@@ -4,6 +4,9 @@ body::-webkit-scrollbar {
     display: none; 
 }
 
+::placeholder {
+    color: rgba(0, 0, 0, 0.5) !important;
+}
 
 a:link {
     color: rgba(255, 255, 0, 1);
