+++ hackru-frontend/src/App.jsx
@@ -1,7 +1,7 @@
 /**
  * @author Shivan Modha
  * @description App object class, which is the first thing to be rendered
- * @version 0.0.1
+ * @version 0.0.2
  * Created 12/09/18
  */
 /***************************************************************IMPORTS***************************************************************/
@@ -10,8 +10,10 @@ import { BrowserRouter, Route, Switch } from "react-router-dom"; // React router
 import {
     LandingPage,
     DashboardPage,
+    LoginPage,
     E404 } from "./components/Pages"; // Router Pages
-import MLHBadge from "./MLHBadge";
+import FlyingLogo from "./FlyingLogo" // The logos that go up through the page
+import MLHBadge from "./MLHBadge"; // We need this to qualify as an official MLH event
 /***************************************************************IMPORTS***************************************************************/
 
 /*****************************************************************APP*****************************************************************/
@@ -29,9 +31,17 @@ class App extends Component {
                 <div>
                     {/* We need to show this on our webpage at all times, so we're just going to dump it in the root */}
                     <MLHBadge />
+                    {/* We put the background here so that even after the page reroutes to different urls, the flying
+                        logos will stay constant, allowing for a seemless user experience. First, we render the logos
+                        then we render the background ontop of them, allowing the logos to fly behind the clouds */}
+                    <div style={{ position: "fixed", zIndex: 1, width: "100%", height: "100%", left: 0, top: 0, opacity: 0.5 }}>
+                        <FlyingLogo url={"./assets/icons/hru-alien-noplat-white.png"} />
+                    </div>
+                    <div style={{ position: "fixed", zIndex: 2, width: "100%", height: "100%", left: 0, top: 0, background: "url(./assets/hru-background-small.png)", backgroundSize: "cover", opacity: 0.5 }}></div>
                     <Switch>
                         {/* This is where the URL routing magic actually happens */}
                         <Route exact path="/" component={LandingPage} />
+                        <Route exact path="/login" component={LoginPage} />
                         <Route exact path="/dashboard" component={DashboardPage} />
                         {/* If none of the other urls were matched, we will show a 404 page to the user */}
                         <Route component={E404} />
