+++ hackru-frontend/src/components/Landing/Sections/Schedule.jsx
@@ -17,17 +17,17 @@ class Schedule extends Component {
         return (
             <div>
 
-                <div class="row">
-                    <div class="col-12">
-                        <h1 class="display-4 theme-font">Schedule</h1>
+                <div className="row">
+                    <div className="col-12">
+                        <h1 className="display-4 theme-font">Schedule</h1>
                     </div>
                 </div>
                 <hr />
 
-                <div class="row row-content-box">
-                    <div class="col-12 col-lg-6">
-                        <h3 class="text-center mb-2">Saturday</h3>
-                        <table class="table table-hover">
+                <div className="row row-content-box">
+                    <div className="col-12 col-lg-6">
+                        <h3 className="text-center mb-2">Saturday</h3>
+                        <table className="table table-hover">
                             <thead>
                                 <tr>
                                     <th scope="col">Time</th>
@@ -65,9 +65,9 @@ class Schedule extends Component {
                         </table>
                     </div>
 
-                    <div class="col-12 mt-5 mt-lg-0 col-lg-6">
-                        <h3 class="text-center mb-2">Sunday</h3>
-                        <table class="table table-hover">
+                    <div className="col-12 mt-5 mt-lg-0 col-lg-6">
+                        <h3 className="text-center mb-2">Sunday</h3>
+                        <table className="table table-hover">
                             <thead>
                                 <tr>
                                     <th scope="col">Time</th>
