+++ hackru-frontend/src/components/Pages.jsx
@@ -7,6 +7,7 @@
 /***************************************************************IMPORTS***************************************************************/
 import LandingPage from "./Landing/Landing"
 import DashboardPage from "./Dashboard"
+import LoginPage from "./Login"
 import E404 from "./Errors/E404"
 /***************************************************************IMPORTS***************************************************************/
 
@@ -14,6 +15,7 @@ import E404 from "./Errors/E404"
 export {
     LandingPage,
     DashboardPage,
+    LoginPage,
     E404
 }
 /***************************************************************EXPORTS***************************************************************/
